+++ lib/Camera.dart
@@ -1,0 +1,49 @@
+// import 'dart:async';
+// import 'package:flutter/material.dart';
+// import 'package:camera/camera.dart';
+
+// List<CameraDescription> cameras;
+
+// Future<void> main() async {
+//   WidgetsFlutterBinding.ensureInitialized();
+
+//   cameras = await availableCameras();
+//   runApp(CameraApp());
+// }
+
+// class CameraApp extends StatefulWidget {
+//   @override
+//   _CameraAppState createState() => _CameraAppState();
+// }
+
+// class _CameraAppState extends State<CameraApp> {
+//   CameraController controller;
+
+//   @override
+//   void initState() {
+//     super.initState();
+//     controller = CameraController(cameras[0], ResolutionPreset.max);
+//     controller.initialize().then((_) {
+//       if (!mounted) {
+//         return;
+//       }
+//       setState(() {});
+//     });
+//   }
+
+//   @override
+//   void dispose() {
+//     controller?.dispose();
+//     super.dispose();
+//   }
+
+//   @override
+//   Widget build(BuildContext context) {
+//     if (!controller.value.isInitialized) {
+//       return Container();
+//     }
+//     return MaterialApp(
+//       home: CameraPreview(controller),
+//     );
+//   }
+// }
