+++ lib/Support.dart
@@ -1,197 +1,197 @@
-import 'package:flutter/material.dart';
+// import 'package:flutter/material.dart';
         
         
-class Iphone13promaxsupportWidget extends StatefulWidget {
-  @override
-  _Iphone13promaxsupportWidgetState createState() => _Iphone13promaxsupportWidgetState();
-}
+// class Iphone13promaxsupportWidget extends StatefulWidget {
+//   @override
+//   _Iphone13promaxsupportWidgetState createState() => _Iphone13promaxsupportWidgetState();
+// }
         
-class _Iphone13promaxsupportWidgetState extends State<Iphone13promaxsupportWidget> {
-  @override
-  Widget build(BuildContext context) {
-  // Figma Flutter Generator Iphone13promaxsupportWidget - FRAME
+// class _Iphone13promaxsupportWidgetState extends State<Iphone13promaxsupportWidget> {
+//   @override
+//   Widget build(BuildContext context) {
+//   // Figma Flutter Generator Iphone13promaxsupportWidget - FRAME
         
-  return Container(
-    width: 428,
-    height: 926,
-    decoration: BoxDecoration(
-      color : Color.fromRGBO(215, 221, 233, 1),
-    ),
-    child: Stack(
-      children: <Widget>[
-        Positioned(
-          top: 72,
-          left: 124,
-          child: AssetImage('assets/images/logo.png') //groccess img
-        ),
-        Positioned(
-          top: 316,
-          left: 39,
-          child: null //contact us method
-        ), Positioned(
-          top: 253,
-          left: 346,
-          child: null // 
-        ),
-        Positioned(
-        top: 253,
-        left: 278,
-        child: null
-      ),Positioned(
-        top: 846,
-        left: -2,
-        child: Container(
-      decoration: BoxDecoration(
-          color : Color.fromRGBO(255, 255, 255, 1),
-  ),
-      padding: EdgeInsets.symmetric(horizontal: 0, vertical: 0),
-      child: Row(
-        mainAxisSize: MainAxisSize.min,
+//   return Container(
+//     width: 428,
+//     height: 926,
+//     decoration: BoxDecoration(
+//       color : Color.fromRGBO(215, 221, 233, 1),
+//     ),
+//     child: Stack(
+//       children: <Widget>[
+//         Positioned(
+//           top: 72,
+//           left: 124,
+//           child: AssetImage('assets/images/logo.png') //groccess img
+//         ),
+//         Positioned(
+//           top: 316,
+//           left: 39,
+//           child: null //contact us method
+//         ), Positioned(
+//           top: 253,
+//           left: 346,
+//           child: null // 
+//         ),
+//         Positioned(
+//         top: 253,
+//         left: 278,
+//         child: null
+//       ),Positioned(
+//         top: 846,
+//         left: -2,
+//         child: Container(
+//       decoration: BoxDecoration(
+//           color : Color.fromRGBO(255, 255, 255, 1),
+//   ),
+//       padding: EdgeInsets.symmetric(horizontal: 0, vertical: 0),
+//       child: Row(
+//         mainAxisSize: MainAxisSize.min,
         
-        children: <Widget>[
-          Container(
-      decoration: BoxDecoration(
-          color : Color.fromRGBO(253, 253, 255, 1),
-  ),
-      padding: EdgeInsets.symmetric(horizontal: 0, vertical: 0),
-      child: Row(
-        mainAxisSize: MainAxisSize.min,
+//         children: <Widget>[
+//           Container(
+//       decoration: BoxDecoration(
+//           color : Color.fromRGBO(253, 253, 255, 1),
+//   ),
+//       padding: EdgeInsets.symmetric(horizontal: 0, vertical: 0),
+//       child: Row(
+//         mainAxisSize: MainAxisSize.min,
         
-        children: <Widget>[
-          Container(
-      decoration: BoxDecoration(
+//         children: <Widget>[
+//           Container(
+//       decoration: BoxDecoration(
           
-  ),
-      padding: EdgeInsets.symmetric(horizontal: 0, vertical: 16),
-      child: Column(
-        mainAxisSize: MainAxisSize.min,
+//   ),
+//       padding: EdgeInsets.symmetric(horizontal: 0, vertical: 16),
+//       child: Column(
+//         mainAxisSize: MainAxisSize.min,
         
-        children: <Widget>[null, SizedBox(height : 8),
-Text('Statistics', textAlign: TextAlign.center, style: TextStyle(
-        color: Color.fromRGBO(73, 69, 79, 1),
-        fontFamily: 'Roboto',
-        fontSize: 12,
-        letterSpacing: 0.5,
-        fontWeight: FontWeight.normal,
-        height: 1.3333333333333333
-      ),),
-],
-      ),
-    ), SizedBox(width : 0),
-Container(
-      decoration: BoxDecoration(
+//         children: <Widget>[null, SizedBox(height : 8),
+// Text('Statistics', textAlign: TextAlign.center, style: TextStyle(
+//         color: Color.fromRGBO(73, 69, 79, 1),
+//         fontFamily: 'Roboto',
+//         fontSize: 12,
+//         letterSpacing: 0.5,
+//         fontWeight: FontWeight.normal,
+//         height: 1.3333333333333333
+//       ),),
+// ],
+//       ),
+//     ), SizedBox(width : 0),
+// Container(
+//       decoration: BoxDecoration(
           
-  ),
-      padding: EdgeInsets.symmetric(horizontal: 0, vertical: 16),
-      child: Column(
-        mainAxisSize: MainAxisSize.min,
+//   ),
+//       padding: EdgeInsets.symmetric(horizontal: 0, vertical: 16),
+//       child: Column(
+//         mainAxisSize: MainAxisSize.min,
         
-        children: <Widget>[null, SizedBox(height : 8),
-Text('Callendar', textAlign: TextAlign.center, style: TextStyle(
-        color: Color.fromRGBO(0, 0, 0, 1),
-        fontFamily: 'Roboto',
-        fontSize: 12,
-        letterSpacing: 0.5,
-        fontWeight: FontWeight.normal,
-        height: 1.3333333333333333
-      ),),
-],
-      ),
-    ), SizedBox(width : 0),
-Container(
-      decoration: BoxDecoration(
+//         children: <Widget>[null, SizedBox(height : 8),
+// Text('Callendar', textAlign: TextAlign.center, style: TextStyle(
+//         color: Color.fromRGBO(0, 0, 0, 1),
+//         fontFamily: 'Roboto',
+//         fontSize: 12,
+//         letterSpacing: 0.5,
+//         fontWeight: FontWeight.normal,
+//         height: 1.3333333333333333
+//       ),),
+// ],
+//       ),
+//     ), SizedBox(width : 0),
+// Container(
+//       decoration: BoxDecoration(
           
-  ),
-      padding: EdgeInsets.symmetric(horizontal: 0, vertical: 16),
-      child: Column(
-        mainAxisSize: MainAxisSize.min,
+//   ),
+//       padding: EdgeInsets.symmetric(horizontal: 0, vertical: 16),
+//       child: Column(
+//         mainAxisSize: MainAxisSize.min,
         
-        children: <Widget>[null, SizedBox(height : 8),
-Text('Camera', textAlign: TextAlign.center, style: TextStyle(
-        color: Color.fromRGBO(0, 0, 0, 1),
-        fontFamily: 'Roboto',
-        fontSize: 12,
-        letterSpacing: 0.5,
-        fontWeight: FontWeight.normal,
-        height: 1.3333333333333333
-      ),),
-],
-      ),
-    ), SizedBox(width : 0),
-Container(
-      decoration: BoxDecoration(
+//         children: <Widget>[null, SizedBox(height : 8),
+// Text('Camera', textAlign: TextAlign.center, style: TextStyle(
+//         color: Color.fromRGBO(0, 0, 0, 1),
+//         fontFamily: 'Roboto',
+//         fontSize: 12,
+//         letterSpacing: 0.5,
+//         fontWeight: FontWeight.normal,
+//         height: 1.3333333333333333
+//       ),),
+// ],
+//       ),
+//     ), SizedBox(width : 0),
+// Container(
+//       decoration: BoxDecoration(
           
-  ),
-      padding: EdgeInsets.symmetric(horizontal: 0, vertical: 16),
-      child: Column(
-        mainAxisSize: MainAxisSize.min,
+//   ),
+//       padding: EdgeInsets.symmetric(horizontal: 0, vertical: 16),
+//       child: Column(
+//         mainAxisSize: MainAxisSize.min,
         
-        children: <Widget>[null, SizedBox(height : 8),
-Text('Profile', textAlign: TextAlign.center, style: TextStyle(
-        color: Color.fromRGBO(0, 0, 0, 1),
-        fontFamily: 'Roboto',
-        fontSize: 12,
-        letterSpacing: 0.5,
-        fontWeight: FontWeight.normal,
-        height: 1.3333333333333333
-      ),),
-],
-      ),
-    ), SizedBox(width : 0),
-Container(
-      decoration: BoxDecoration(
+//         children: <Widget>[null, SizedBox(height : 8),
+// Text('Profile', textAlign: TextAlign.center, style: TextStyle(
+//         color: Color.fromRGBO(0, 0, 0, 1),
+//         fontFamily: 'Roboto',
+//         fontSize: 12,
+//         letterSpacing: 0.5,
+//         fontWeight: FontWeight.normal,
+//         height: 1.3333333333333333
+//       ),),
+// ],
+//       ),
+//     ), SizedBox(width : 0),
+// Container(
+//       decoration: BoxDecoration(
           
-  ),
-      padding: EdgeInsets.symmetric(horizontal: 0, vertical: 12),
-      child: Column(
-        mainAxisSize: MainAxisSize.min,
+//   ),
+//       padding: EdgeInsets.symmetric(horizontal: 0, vertical: 12),
+//       child: Column(
+//         mainAxisSize: MainAxisSize.min,
         
-        children: <Widget>[Container(
-      width: 64,
-      height: 32,
-      decoration: BoxDecoration(
-          color : Color.fromRGBO(255, 255, 255, 1),
-  ),
-      child: Stack(
-        children: <Widget>[
-          Positioned(
-        top: 0,
-        left: 0,
-        child: SvgPicture.asset(
-        'assets/images/activeindicator.svg',
-        semanticsLabel: 'activeindicator'
-      );
-      ),Positioned(
-        top: 4,
-        left: 20,
-        child: null
-      ),
-        ]
-      )
-    ), SizedBox(height : 8),
-Text('Support', textAlign: TextAlign.center, style: TextStyle(
-        color: Color.fromRGBO(29, 25, 43, 1),
-        fontFamily: 'Roboto',
-        fontSize: 12,
-        letterSpacing: 0.5,
-        fontWeight: FontWeight.normal,
-        height: 1.3333333333333333
-      ),),
-],
-      ),
-    ),
+//         children: <Widget>[Container(
+//       width: 64,
+//       height: 32,
+//       decoration: BoxDecoration(
+//           color : Color.fromRGBO(255, 255, 255, 1),
+//   ),
+//       child: Stack(
+//         children: <Widget>[
+//           Positioned(
+//         top: 0,
+//         left: 0,
+//         child: SvgPicture.asset(
+//         'assets/images/activeindicator.svg',
+//         semanticsLabel: 'activeindicator'
+//       );
+//       ),Positioned(
+//         top: 4,
+//         left: 20,
+//         child: null
+//       ),
+//         ]
+//       )
+//     ), SizedBox(height : 8),
+// Text('Support', textAlign: TextAlign.center, style: TextStyle(
+//         color: Color.fromRGBO(29, 25, 43, 1),
+//         fontFamily: 'Roboto',
+//         fontSize: 12,
+//         letterSpacing: 0.5,
+//         fontWeight: FontWeight.normal,
+//         height: 1.3333333333333333
+//       ),),
+// ],
+//       ),
+//     ),
 
-        ],
-      ),
-    ),
+//         ],
+//       ),
+//     ),
 
-        ],
-      ),
-    )
-      ),
-        ]
-      )
-    );
-        }
-        }
+//         ],
+//       ),
+//     )
+//       ),
+//         ]
+//       )
+//     );
+//         }
+//         }
