+++ lib/bottomNavBar.dart
@@ -1,0 +1,42 @@
+import 'package:flutter/material.dart';
+
+class BottomMenu extends StatelessWidget {
+  final selectedIndex;
+  ValueChanged<int> onClicked;
+  BottomMenu({this.selectedIndex, required this.onClicked});
+  @override
+  Widget build(BuildContext context) {
+    return BottomNavigationBar(
+      items: const <BottomNavigationBarItem>[
+        BottomNavigationBarItem(
+          icon: Icon(Icons.shopping_cart),
+          label: 'Statistics',
+          backgroundColor: Colors.grey,
+        ),
+        BottomNavigationBarItem(
+          icon: Icon(Icons.calendar_today_rounded),
+          label: 'Calendar',
+          backgroundColor: Colors.grey,
+        ),
+        BottomNavigationBarItem(
+          icon: Icon(Icons.camera),
+          label: 'Camera',
+          backgroundColor: Colors.grey,
+        ),
+        BottomNavigationBarItem(
+          icon: Icon(Icons.face),
+          label: 'Profile',
+          backgroundColor: Colors.grey,
+        ),
+        BottomNavigationBarItem(
+          icon: Icon(Icons.support_agent),
+          label: 'Support',
+          backgroundColor: Colors.grey,
+        ),
+      ],
+      currentIndex: selectedIndex,
+      selectedItemColor: Color.fromARGB(255, 140, 187, 241),
+      onTap: onClicked,
+    );
+  }
+}
