+++ lib/main.dart
@@ -1,18 +1,21 @@
+import 'package:first_app_figma/bottomNavBar.dart';
 import 'package:flutter/material.dart';
-import 'dart:async';
-import 'package:camera/camera.dart';
+import 'Profile.dart';
 
-late List<CameraDescription> cameras = [];
+// import 'dart:async';
+// import 'package:camera/camera.dart';
 
-Future<void> main() async {
-  WidgetsFlutterBinding.ensureInitialized();
+// late List<CameraDescription> cameras = [];
 
-  WidgetsFlutterBinding.ensureInitialized();
-  try {
-    cameras = await availableCameras();
-  } on CameraException catch (e) {
-    logError(e.code, e.description);
-  }
+void main() {
+  // WidgetsFlutterBinding.ensureInitialized();
+
+  // WidgetsFlutterBinding.ensureInitialized();
+  // try {
+  //   cameras = await availableCameras();
+  // } on CameraException catch (e) {
+  //   logError(e.code, e.description);
+  // }
   runApp(MyApp());
 }
 
@@ -38,30 +41,18 @@ class MyStatefulWidget extends StatefulWidget {
 }
 
 class _MyStatefulWidgetState extends State<MyStatefulWidget> {
-  late CameraController controller;
-  @override
-  void initState() {
-    super.initState();
-    controller = CameraController(cameras[0], ResolutionPreset.max);
-    controller.initialize().then((_) {
-      if (!mounted) {
-        return;
-      }
-      setState(() {});
-    });
-  }
-
-  @override
-  void dispose() {
-    controller?.dispose();
-    super.dispose();
-  }
-
   int _selectedIndex = 0;
   static const TextStyle optionStyle = TextStyle(
     fontSize: 12,
     fontFamily: 'Roboto',
   );
+  List screens = [
+    profile(widgetOptions: _widgetOptions, selectedIndex: 0),
+    profile(widgetOptions: _widgetOptions, selectedIndex: 1),
+    profile(widgetOptions: _widgetOptions, selectedIndex: 2),
+    profile(widgetOptions: _widgetOptions, selectedIndex: 3),
+    profile(widgetOptions: _widgetOptions, selectedIndex: 4)
+  ];
   static const List<Widget> _widgetOptions = <Widget>[
     Text(
       'Index 0: Statistics',
@@ -90,8 +81,6 @@ class _MyStatefulWidgetState extends State<MyStatefulWidget> {
       _selectedIndex = index;
     });
     if (index == 2) {
-      Navigator.of(context).push(
-          MaterialPageRoute(builder: (context) => CameraPreview(controller)));
       print("i am the camera tab");
     }
   }
@@ -100,66 +89,11 @@ class _MyStatefulWidgetState extends State<MyStatefulWidget> {
   Widget build(BuildContext context) {
     return Scaffold(
       backgroundColor: Color.fromARGB(255, 215, 221, 233),
-      body: Column(children: [
-        SizedBox(height: 40),
-        Center(
-          child: _widgetOptions.elementAt(_selectedIndex),
-        ),
-        Padding(
-          padding: const EdgeInsets.all(8.0),
-          child: Container(
-              height: 136.0,
-              width: 179.0,
-              decoration: const BoxDecoration(
-                borderRadius: BorderRadius.only(
-                  topLeft: Radius.circular(23),
-                  topRight: Radius.circular(23),
-                  bottomLeft: Radius.circular(23),
-                  bottomRight: Radius.circular(23),
-                ),
-                image: DecorationImage(
-                    fit: BoxFit.fill,
-                    image: AssetImage('assets/images/logo.png')),
-                color: Color.fromRGBO(253, 253, 255, 1),
-              )),
-        )
-      ]),
-      bottomNavigationBar: NavBar(),
-    );
-  }
-
-  BottomNavigationBar NavBar() {
-    return BottomNavigationBar(
-      items: const <BottomNavigationBarItem>[
-        BottomNavigationBarItem(
-          icon: Icon(Icons.shopping_cart),
-          label: 'Statistics',
-          backgroundColor: Colors.grey,
-        ),
-        BottomNavigationBarItem(
-          icon: Icon(Icons.calendar_today_rounded),
-          label: 'Calendar',
-          backgroundColor: Colors.grey,
-        ),
-        BottomNavigationBarItem(
-          icon: Icon(Icons.camera),
-          label: 'Camera',
-          backgroundColor: Colors.grey,
-        ),
-        BottomNavigationBarItem(
-          icon: Icon(Icons.face),
-          label: 'Profile',
-          backgroundColor: Colors.grey,
-        ),
-        BottomNavigationBarItem(
-          icon: Icon(Icons.support_agent),
-          label: 'Support',
-          backgroundColor: Colors.grey,
-        ),
-      ],
-      currentIndex: _selectedIndex,
-      selectedItemColor: Color.fromARGB(255, 140, 187, 241),
-      onTap: _onItemTapped,
+      body: screens.elementAt(_selectedIndex),
+      bottomNavigationBar: BottomMenu(
+        selectedIndex: _selectedIndex,
+        onClicked: _onItemTapped,
+      ),
     );
   }
 }
