+++ pubspec.yaml
@@ -36,7 +36,7 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
-  camera: ^0.9.4+5
+  # camera: ^0.9.4+5
 dev_dependencies:
   flutter_test:
     sdk: flutter
